WT-18258 Run operations during async step-down window. - refactor, names, visibility
